modify uiux and update db url for heroku

+++ client/components/document/DocumentList.js
@@ -12,7 +12,7 @@ class DocumentList extends React.Component {
   }
 
   componentDidMount() {
-    $('.editDoc').click(this.editDocument);
+    $('.tooltipped').tooltip({ delay: 50 });
   }
   editDocument(e) {
     e.preventDefault();
@@ -22,8 +22,11 @@ class DocumentList extends React.Component {
   deleteDocument(e) {
     e.preventDefault();
     const documentId = e.target.id;
-    this.props.actions.deleteDocument(documentId)
-    .then(() => toastr.success('Document Successfully Deleted'));
+    const result = confirm('Do you want to delete this docuement?');
+    if (result) {
+      this.props.actions.deleteDocument(documentId)
+      .then(() => toastr.success('Document Successfully Deleted'));
+    }
   }
   render() {
     return (
@@ -35,25 +38,35 @@ class DocumentList extends React.Component {
           .map(document => <div id="card-alert" className="card white"
           key={document.id}>
             <div className="card-content pink-text">
-              <a className="pointer"
+              <a className="pointer tooltipped"
+                data-position="bottom" data-delay="50"
+                data-tooltip="click on me to view"
                 href="#modal1" id={document.id}
                 onClick={this.editDocument}>
-              {document.title}
+              Title: {document.title}
               </a>
             </div>
             <div className="fixed-action-btn horizontal click-to-toggle edit">
-              <a className="btn-floating pink">
+              <a className="btn-floating pink tooltipped"
+                data-position="top" data-delay="50"
+                data-tooltip="click to view more"
+                >
                 <i className="material-icons">more_vert</i>
               </a>
               <ul>
                 <li onClick={this.editDocument} className="editDoc">
-                  <a title="edit" href="#modal1"
-                  className="btn-floating pink">
+                  <a href="#modal1"
+                  className="btn-floating pink tooltipped"
+                  data-position="bottom" data-delay="50"
+                  data-tooltip="edit document">
                     <i id={document.id} className="material-icons">mode_edit</i>
                   </a>
                 </li>
                 <li onClick={this.deleteDocument}>
-                  <a title="delete" className="btn-floating red darken-1">
+                  <a className="btn-floating red darken-1 tooltipped"
+                    data-position="bottom" data-delay="50"
+                    data-tooltip="delete document"
+                    >
                     <i id={document.id} className="material-icons">delete</i>
                   </a>
                 </li>
@@ -71,7 +84,7 @@ DocumentList.propsTypes = {
 
 /**
  *
- *
+ * dispatch document actions
  * @param {any} dispatch
  * @returns {any}
  */
